refactor(frontend): encapsulate createApiClient behind two-client API
The codebase already documents a two-client surface — the default
browser Proxy from $lib/api/client and createServerClient from
$lib/api/server. createApiClient was a leaky implementation detail
on the public surface, and #420 was the kind of bug that leaky
factories enable. After commit 1 dropped the external caller count
to zero, this commit moves the factory to $lib/api/internal/, splits
the cookie helper out to $lib/api/csrf, and adds an ESLint rule
banning $lib/api/internal/* imports from app code.
